Add files via upload
I have created a simple portfolio website using FLASK FRAMEWORK . 
The main.css file includes css  coding and we have linked it with the layout.html file. Th emain.css file is in static folder. Python looks for static files in static files in static folder.
The home.html,about.html and layout.html are included in the template folder.
The main Flask app(code for the website) is in the tut1.py file.